Create functionality to display the DOM data on live site
