+++ src/scripts.js
@@ -6,6 +6,8 @@ import Hydration from './Hydration';
 import Activity from './Activity';
 import './css/styles.css';
 import displayCharts from './charts.js';
+import domUpdates from './domUpdates.js';
+import {displayUserInfoCard, displayUserData, displayAvgStepGoal, displaySleepQuality} from './domUpdates.js';
 import './images/add.png';
 import './images/add-hover.png';
 let sleepData;
@@ -18,6 +20,8 @@ let userData;
 let userh20;
 let allUsersSleep;
 const todaysDate = '2020/01/22';
+
+domUpdates()
 
 import {
   fetchUserData,
@@ -44,130 +48,115 @@ const allData = (user, sleep, activity, h20Data) => {
   currentUser = new User(userRepository.users[getRandomIndex(userRepository.users)]);
   allUsersSleep = new Sleep(sleepData)
   userh20 = new Hydration(h20Data);
-  displayUserInfoCard();
+  displayUserInfoCard(currentUser);
+  displayUserData(currentUser);
+  displayAvgStepGoal(userRepository);
+  displaySleepQuality(allUsersSleep, currentUser, todaysDate);
   displayCharts(userh20, allUsersSleep, currentUser, todaysDate);
 };
 
-//QUERY SELECTORS
-const greeting = document.getElementById('firstName');
-const userSteps = document.getElementById('userSteps');
-const userWater = document.getElementById('userWater');
-const userSleep = document.getElementById('userSleep');
-const userActivity = document.getElementById('userActivity');
-const userEmail = document.getElementById('userEmail');
-const userAddress = document.getElementById('userAddress');
-const userStride = document.getElementById('userStride');
-const userStepGoal = document.getElementById('userStepGoal');
-const avgStepGoal = document.getElementById('avgStepGoal');
-const dailySleepQuality = document.getElementById('dailySleepQuality');
-const avgSleepQuality = document.getElementById('avgSleepQuality');
 
-//Form Field Query Selectors & Event Listeners
-const logWater = document.getElementById('logWater');
-const waterChart = document.getElementById('waterChart');
-const waterForm = document.getElementById('waterForm');
-const logSleep = document.getElementById('logSleep');
-const sleepChart = document.getElementById('sleepChart');
-const sleepForm = document.getElementById('sleepForm');
-const logActivity = document.getElementById('logActivity');
-const activityReport = document.getElementById('activityReport');
-const activityForm = document.getElementById('activityForm');
+//
+// //QUERY SELECTORS
+// const greeting = document.getElementById('firstName');
+// const userSteps = document.getElementById('userSteps');
+// const userWater = document.getElementById('userWater');
+// const userSleep = document.getElementById('userSleep');
+// const userActivity = document.getElementById('userActivity');
+// const userEmail = document.getElementById('userEmail');
+// const userAddress = document.getElementById('userAddress');
+// const userStride = document.getElementById('userStride');
+// const userStepGoal = document.getElementById('userStepGoal');
+// const avgStepGoal = document.getElementById('avgStepGoal');
+// const dailySleepQuality = document.getElementById('dailySleepQuality');
+// const avgSleepQuality = document.getElementById('avgSleepQuality');
+//
+// //Form Field Query Selectors & Event Listeners
+// const logWater = document.getElementById('logWater');
+// const waterChart = document.getElementById('waterChart');
+// const waterForm = document.getElementById('waterForm');
+// const logSleep = document.getElementById('logSleep');
+// const sleepChart = document.getElementById('sleepChart');
+// const sleepForm = document.getElementById('sleepForm');
+// const logActivity = document.getElementById('logActivity');
+// const activityReport = document.getElementById('activityReport');
+// const activityForm = document.getElementById('activityForm');
+//
+// logWater.addEventListener('click', showWaterForm);
+// logSleep.addEventListener('click', showSleepForm);
+// logActivity.addEventListener('click', showActivityForm);
 
-logWater.addEventListener('click', showWaterForm);
-logSleep.addEventListener('click', showSleepForm);
-logActivity.addEventListener('click', showActivityForm);
-
-//EVENT LISTENERS
+// //EVENT LISTENERS
 window.addEventListener('load', () => {
   fetchAllData();
 });
-
-//HELPER FUNCTIONS
-function hide(element) {
-  element.classList.add('hidden')
-}
-
-function show(element) {
-  element.classList.remove('hidden')
-}
-
-function toggle(element) {
-  element.classList.toggle('hidden');
-}
-
+//
+// //HELPER FUNCTIONS
+// function hide(element) {
+//   element.classList.add('hidden')
+// }
+//
+// function show(element) {
+//   element.classList.remove('hidden')
+// }
+//
+// function toggle(element) {
+//   element.classList.toggle('hidden');
+// }
+//
 const getRandomIndex = (array) => {
   return Math.floor(Math.random() * array.length)
 };
-
-//DATA DISPLAY FUNCTIONS
-
-const displayUserInfoCard = () => {
-  greeting.innerText = `Welcome, ${currentUser.getFirstName()}!`;
-  displayUserData();
-  // displayUserEmail();
-  // displayUserAddress();
-  // displayUserStride();
-  // displayUserDailyStepGoal();
-  displayAvgStepGoal();
-  displaySleepQuality();
-};
-
-const displayUserData = () => {
-  userEmail.innerText = `${currentUser.email}`;
-  userAddress.innerText = `${currentUser.address}`;
-  userStride.innerText = `${currentUser.strideLength}`;
-  userStepGoal.innerText = `${currentUser.dailyStepGoal}`;
-};
-
-// const displayUserEmail = () => {
+//
+// //DATA DISPLAY FUNCTIONS
+//
+// const displayUserInfoCard = () => {
+//   greeting.innerText = `Welcome, ${currentUser.getFirstName()}!`;
+//   displayUserData();
+//   displayAvgStepGoal();
+//   displaySleepQuality();
+// };
+//
+// const displayUserData = () => {
 //   userEmail.innerText = `${currentUser.email}`;
-// };
-//
-// const displayUserAddress = () => {
 //   userAddress.innerText = `${currentUser.address}`;
-// };
-//
-// const displayUserStride = () => {
 //   userStride.innerText = `${currentUser.strideLength}`;
-// };
-//
-// const displayUserDailyStepGoal = () => {
 //   userStepGoal.innerText = `${currentUser.dailyStepGoal}`;
 // };
-
-const displayAvgStepGoal = () => {
-  avgStepGoal.innerText = `${userRepository.getAvgStepCount()}`;
-};
-
-const displaySleepQuality = () => {
-  dailySleepQuality.innerText = allUsersSleep.getDailySleepQual(currentUser.id, todaysDate);
-  avgSleepQuality.innerText = allUsersSleep.getAvgDailySleepQual(currentUser.id);
-};
-
-//FUNCTIONS FOR USER AVERAGES
-  /*** FYI - pulled the two below from the charts we're removing ***/
-
-// let userAvgSleepHours = allUsersSleep.getAvgSleepPerDay(currentUser.id);
-
-// let userAvgWater = userh20.getAvgOuncesPerDay(currentUser.id);
-
-
-
-//FUNCTIONS FOR COMMUNITY AVERAGES
-
-//FORM FUNCTIONS
-
-function showWaterForm() {
-  toggle(waterChart);
-  toggle(waterForm);
-};
-
-function showSleepForm() {
-  toggle(sleepChart);
-  toggle(sleepForm);
-};
-
-function showActivityForm() {
-  toggle(activityReport);
-  toggle(activityForm);
-};
+//
+// const displayAvgStepGoal = () => {
+//   avgStepGoal.innerText = `${userRepository.getAvgStepCount()}`;
+// };
+//
+// const displaySleepQuality = () => {
+//   dailySleepQuality.innerText = allUsersSleep.getDailySleepQual(currentUser.id, todaysDate);
+//   avgSleepQuality.innerText = allUsersSleep.getAvgDailySleepQual(currentUser.id);
+// };
+//
+// //FUNCTIONS FOR USER AVERAGES
+//   /*** FYI - pulled the two below from the charts we're removing ***/
+//
+// // let userAvgSleepHours = allUsersSleep.getAvgSleepPerDay(currentUser.id);
+// //
+// // let userAvgWater = userh20.getAvgOuncesPerDay(currentUser.id);
+//
+//
+//
+// //FUNCTIONS FOR COMMUNITY AVERAGES
+//
+// //FORM FUNCTIONS
+//
+// function showWaterForm() {
+//   toggle(waterChart);
+//   toggle(waterForm);
+// };
+//
+// function showSleepForm() {
+//   toggle(sleepChart);
+//   toggle(sleepForm);
+// };
+//
+// function showActivityForm() {
+//   toggle(activityReport);
+//   toggle(activityForm);
+// };
